Changes on 08032025

+++ src/main/java/com/MovieMatrix/services/MovieService.java
@@ -1,33 +1,57 @@
 package com.MovieMatrix.services;
 
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.MovieMatrix.models.Movie;
 import com.MovieMatrix.repositories.MovieRepository;
-
+import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MovieService {
+  private final MovieRepository movieRepository;
 
-  @Autowired
-  private MovieRepository movieRepository;
+  public MovieService(MovieRepository movieRepository) {
+    this.movieRepository = movieRepository;
+  }
 
   public List<Movie> getAllMovies() {
     return movieRepository.findAll();
   }
 
-  public Movie getMovieById(Long id) {
-    return movieRepository.findById(id).orElse(null);
+  public Optional<Movie> getMovieById(Long id) {
+    return movieRepository.findById(id);
+  }
+
+  public Optional<Movie> getMovieByTitle(String title) {
+    return movieRepository.findByTitle(title);
   }
 
   public Movie addMovie(Movie movie) {
     return movieRepository.save(movie);
   }
 
-  public void deleteMovie(Long id) {
-    movieRepository.deleteById(id);
+  public Movie updateMovie(Long id, Movie movieDetails) {
+    return movieRepository.findById(id).map(movie -> {
+      movie.setTitle(movieDetails.getTitle());
+      movie.setGenre(movieDetails.getGenre());
+      movie.setDirector(movieDetails.getDirector());
+      // movie.setDescription(movieDetails.getDescription());
+      movie.setActors(movieDetails.getActors());
+      movie.setReleaseDate(movieDetails.getReleaseDate());
+      movie.setPosterUrl(movieDetails.getPosterUrl());
+      return movieRepository.save(movie);
+    }).orElseThrow(() -> new RuntimeException("Movie not found"));
+  }
+
+  public Movie saveMovie(Movie movie) {
+    return movieRepository.save(movie);
+  }
+
+  public boolean deleteMovie(Long id) {
+    if (movieRepository.existsById(id)) {
+      movieRepository.deleteById(id);
+      return true;
+    }
+    return false;
   }
 }
